pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.24.0",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "requests>=2.32.4",
    "spacy>=3.8.7",
    "streamlit>=1.46.1",
    "trafilatura>=2.0.0",
]
